refactor(e2e): split shipped-map guards into their own module (ENG-10026)
CodeScene has been failing since d0a4459 — I did not re-check it after the
first round, so it went unnoticed for three commits. The finding is real and
mine: tests/e2e/test_evidence_emitter.py scored 7.98 (below the 9.00 gate)
with Low Cohesion flagged as a critical rule.

It is an accurate read. The file had grown two unrelated halves: tests that
drive the plugin against synthetic maps and simulated results via pytester,
and guards that shell out to a real `pytest --collect-only` to validate the
shipped capability_map.yaml against the repo's actual test collection.
Different subject, different inputs, no shared state or helpers — which is
exactly the LCOM4 split CodeScene names.

Moved to tests/e2e/test_capability_map.py: _collect_repo_nodeids,
_stale_globs, _unmatched_pattern_problems, and the two shipped-map tests.
No behavior change; the guard was re-mutation-tested after the move (a
renamed exclude glob still fails it). Cross-reference added between the two
modules so the split is discoverable.

Verification: 46 tests pass across both files (44 + 2); full
-m "not integration and not live and not e2e" sweep 3366 passed, 2 skipped
(pre-existing env gates); ruff + black + isort clean; radon max CC 7 in the
new module (cap 9).
